refactor: `services` 패키지 내부를 다시 도메인에 따라 `product`, `user`로 분리함
