+++ app/src/test/java/com/codesoom/assignment/services/UserRegisterServiceTest.java
@@ -1,6 +1,7 @@
 package com.codesoom.assignment.services;
 
 import com.codesoom.assignment.domain.User;
+import com.codesoom.assignment.services.user.UserRegisterService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
